+++ silo-core/contracts/hooks/defaulting/DefaultingSiloLogic.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.28;
 
 import {ISilo} from "silo-core/contracts/interfaces/ISilo.sol";
 import {Math} from "openzeppelin5/utils/math/Math.sol";
-import {SafeCast} from "openzeppelin5/utils/math/SafeCast.sol";
+import {SafeCast} from "openzeppelin-contracts/contracts/utils/math/SafeCast.sol";
 
 import {SiloStorageLib} from "silo-core/contracts/lib/SiloStorageLib.sol";
 import {DefaultingRepayLib} from "silo-core/contracts/hooks/defaulting/DefaultingRepayLib.sol";
@@ -44,7 +44,7 @@ contract DefaultingSiloLogic {
         uint256 totalCollateralAssets = $.totalAssets[ISilo.AssetType.Collateral];
 
         // if underflow happens, $.totalAssets[ISilo.AssetType.Collateral] is set to 0 and success is false
-        (success, $.totalAssets[ISilo.AssetType.Collateral]) = totalCollateralAssets.trySub(_assetsToRepay);
+        (success, $.totalAssets[ISilo.AssetType.Collateral]) = Math.trySub(totalCollateralAssets, _assetsToRepay);
         uint256 deductedFromCollateral = _assetsToRepay;
         
         if (!success) {